Move EMFDependsOnPostProcessor into autoconfigure.orm.jpa
Closes gh-24452
